// src/api/hooks/queries/query-key.ts
export enum authQueryKeys {}
export enum appQueryKeys {
  GET_CATEGORY = "get-category",
  GET_CAR_BRANDS = "get-car-brands",
  GET_CAR_TYPES = "get-car-types",
  GET_SERVICE_STATUS = "get-service-status",
  GET_ALL_SERVICES = "get-all-services",
  VIEW_SERVICE = "view-service",
  GET_SETTINGS = "get-settings",
  GET_USER_NOTIFICATIONS = "get-user-notifications",
  GET_USER_WISHLIST = "get-user-wishlist",
  GET_RECENTLY_VIEWED = "get-recently-viewed",
  GET_USER_REFERRAL = "get-user-referral",
  GET_USER_REFERRAL_REWARD_HISTORY = "get-user-referral-reward-history",
  GET_ALL_USER_CHATS = "get-all-user-chats",
  GET_USER_SERVICE_MESSAGES = "get-user-service-messages",
  GET_TERMS_AND_CONDITIONS = "get-terms-and-conditions",
}